backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
supabase>=2.10.0
firecrawl-py>=2.0.0
google-generativeai>=0.8.0
httpx>=0.27.0
pypdf>=5.0.0
apify-client>=1.8.0
twilio>=9.0.0
yt-dlp>=2024.10.7
youtube-transcript-api>=0.6.2
reportlab>=4.0.0
